Fix some more clang-tidy warnings.
Add `const` to a few ret variables.

Add `L`, `U`, and/or `UL` to some constants to eliminate warnings.

.clang-tidy: Disable misc-include-cleaner. It throws a lot of false
positive warnings due to the use of stdafx.h.

[librpcpu] cpuflags_x86.h: Explicitly cast return values to int.

src/librpfile/../librpcpu/cpuflags_x86.h:88:9: warning: narrowing
conversion from 'uint32_t' (aka 'unsigned int') to signed type 'int' is
implementation-defined [bugprone-narrowing-conversions]
   88 |         return (RP_CPU_Flags & RP_CPUFLAG_X86_SSSE3);
      |                ^

[gtk] PIMGTYPE.hpp: Move rp_image_load_png_from_gresource() out of the
extern "C" block, since it returns a C++ shared_ptr<>.

src/gtk/xfce/../PIMGTYPE.hpp:269:28: error: 'rp_image_load_png_from_gresource'
has C-linkage specified, but returns user-defined type 'LibRpTexture::rp_image_ptr'
(aka 'shared_ptr<rp_image>') which is incompatible with C [clang-diagnostic-return-type-c-linkage]
  269 | LibRpTexture::rp_image_ptr rp_image_load_png_from_gresource(const char *filename);
      |                            ^

[librptexture] DidjTex: Handle width and height as int when calling
ImageDecoder functions.

src/librptexture/fileformat/DidjTex.cpp:224:40: warning: narrowing
conversion from 'unsigned int' to signed type 'int' is
implementation-defined [bugprone-narrowing-conversions]
  224 |                                 ImageDecoder::PixelFormat::RGB565, width, height,
      |                                                                    ^
src/librptexture/fileformat/DidjTex.cpp:224:47: warning: narrowing
conversion from 'unsigned int' to signed type 'int' is
implementation-defined [bugprone-narrowing-conversions]
  224 |                                 ImageDecoder::PixelFormat::RGB565, width, height,
      |                                                                           ^

[libromdata] N3DSVerifyKeys: Deleted ctor/dtor should be public.

src/libromdata/tests/../crypto/N3DSVerifyKeys.hpp:54:3: warning: deleted
member function should be public [modernize-use-equals-delete]
   54 |                 N3DSVerifyKeys() = delete;
      |                 ^
src/libromdata/tests/../crypto/N3DSVerifyKeys.hpp:55:3: warning: deleted
member function should be public [modernize-use-equals-delete]
   55 |                 ~N3DSVerifyKeys() = delete;
      |                 ^

+++ src/librpcpu/cpuflags_x86.h
@@ -55,7 +55,7 @@ static FORCEINLINE int RP_CPU_HasMMX(void)
 	if (unlikely(!RP_CPU_Flags_Init)) {
 		RP_CPU_InitCPUFlags();
 	}
-	return (RP_CPU_Flags & RP_CPUFLAG_X86_MMX);
+	return (int)(RP_CPU_Flags & RP_CPUFLAG_X86_MMX);
 #endif
 }
 
@@ -72,7 +72,7 @@ static FORCEINLINE int RP_CPU_HasSSE2(void)
 	if (unlikely(!RP_CPU_Flags_Init)) {
 		RP_CPU_InitCPUFlags();
 	}
-	return (RP_CPU_Flags & RP_CPUFLAG_X86_SSE2);
+	return (int)(RP_CPU_Flags & RP_CPUFLAG_X86_SSE2);
 #endif
 }
 
@@ -85,7 +85,7 @@ static FORCEINLINE int RP_CPU_HasSSSE3(void)
 	if (unlikely(!RP_CPU_Flags_Init)) {
 		RP_CPU_InitCPUFlags();
 	}
-	return (RP_CPU_Flags & RP_CPUFLAG_X86_SSSE3);
+	return (int)(RP_CPU_Flags & RP_CPUFLAG_X86_SSSE3);
 }
 
 /**
@@ -97,7 +97,7 @@ static FORCEINLINE int RP_CPU_HasSSE41(void)
 	if (unlikely(!RP_CPU_Flags_Init)) {
 		RP_CPU_InitCPUFlags();
 	}
-	return (RP_CPU_Flags & RP_CPUFLAG_X86_SSE41);
+	return (int)(RP_CPU_Flags & RP_CPUFLAG_X86_SSE41);
 }
 
 #ifdef __cplusplus
